third_party/WebKit/LayoutTests/fast/table/layout-section-when-table-height-changes-2.html

<!DOCTYPE html>
<p>crbug.com/691968: There should be a blue square below.</p>
<table id="table" style="border-spacing:0; width:100px; height:200px; background:blue;" data-expected-height=100>
    <tr>
        <td></td>
    </tr>
</table>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
    document.body.offsetTop;
    document.getElementById("table").style.height = "100px";
    checkLayout('#table');
</script>
